Feature(SSK-93): remove resources sign retention
