+++ lib/SWServices/StampRetention/StampRetentionService.js
@@ -1,5 +1,5 @@
 const Services = require('../Services.js');
-const StampRequest = require('../Stamp/StampRequest.js');
+const RequestHelper = require('../Helper/RequestHelper.js');
 const Authentication = require('../Authentication/AuthenticationService.js');
 
 /**
@@ -57,7 +57,7 @@ class StampRetentionService extends Services {
 	}
 
 	_sendRequest(xml, version, callback, isB64) {
-		StampRequest.sendReq(this.get_url(), this.get_token(), xml, version, isB64, callback);
+		RequestHelper.sendRequest(this.get_url(), this.get_token(), xml, 'retencion', version, isB64, callback);
 	}
 }
